Fixed Create Queries and Created Errors for missing .env dependencies

+++ backend/db/db.js
@@ -5,6 +5,11 @@ const username = process.env.MYSQL_USERNAME;
 const password = process.env.MYSQL_PASSWORD;
 const database = process.env.MYSQL_DB;
 
+if (!host) throw new Error("Missing MYSQL_HOST in .env");
+if (!username) throw new Error("Missing MYSQL_USERNAME in .env");
+if (!password) throw new Error("Missing MYSQL_PASSWORD in .env");
+if (!database) throw new Error("Missing MYSQL_DB in .env");
+
 let dropQueries = [
     `DROP TABLE IF EXISTS users`,
     `DROP PROCEDURE IF EXISTS get_user_by_first_name`,
@@ -23,34 +28,25 @@ let createQueries = [
         PRIMARY KEY (\`pid\`)
         ) ENGINE=InnoDB AUTO_INCREMENT=2 DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_0900_ai_ci;
         `,
-    `DELIMITER $$
-        CREATE DEFINER=\`root\`@\`localhost\` PROCEDURE \`get_user_by_first_name\`(IN firstName varchar(32))
+    `CREATE PROCEDURE \`get_user_by_first_name\`(IN firstName varchar(32))
         BEGIN
             SELECT * FROM users WHERE firstName = firstName;
-        END$$
-        DELIMITER;`,
-    `DELIMITER $$
-        CREATE DEFINER=\`root\`@\`localhost\` PROCEDURE \`get_user_by_username\`(IN username varchar(32))
+        END;`,
+    `CREATE PROCEDURE \`get_user_by_username\`(IN username varchar(32))
         BEGIN
             SELECT * FROM users WHERE username = username;
-        END$$
-        DELIMITER ;`,
-    `DELIMITER $$
-        CREATE DEFINER=\`root\`@\`localhost\` PROCEDURE \`login\`(IN username varchar(32), password char(60))
+        END;`,
+    `CREATE PROCEDURE \`login\`(IN username varchar(32), password char(60))
         BEGIN
             SELECT * FROM users WHERE username = username AND password = password;
-        END$$
-        DELIMITER ;`,
-    `DELIMITER $$
-        CREATE DEFINER=\`root\`@\`localhost\` PROCEDURE \`register\`(IN firstName varchar(32), lastName varchar(32), username varchar(32), password char(60))
+        END;`,
+    `CREATE PROCEDURE \`register\`(IN firstName varchar(32), lastName varchar(32), username varchar(32), password char(60))
         BEGIN
-            INSERT INTO users (firstName, lastName, username, password) VALUES (firstName, lastName, username, password);
-        END$$
-        DELIMITER ;`,
-]
+            insert into users (firstName, lastName, username, password) values (firstName, lastName, username, password);
+        END;`
+];
 
 //Create a connection to the SQL Database
-
 const dbPool = mysql.createPool({
     host: host,
     user: username,
@@ -59,12 +55,26 @@ const dbPool = mysql.createPool({
 });
 dbPool.query("SELECT * FROM users", (err, rows) => {
     if(err) {
-        dropQueries.forEach((query, index) => {
-            dbPool.query(query);
+        dbPool.getConnection(function(err, con) {
+            if (err) {
+                console.log("[REMAKE] Connection Error: ", err);
+            }
+            dropQueries.forEach((query, index) => {
+                con.query(query, function(err, results) {
+                    if (err) {
+                        console.log("[REMAKE] Drop Query Error: ", err)
+                    }
+                });
+            });
+            createQueries.forEach((query, index) => {
+                con.query(query, function(err, results) {
+                    console.log(query);
+                    if (err) {
+                        console.log("[REMAKE] Create Query Error: ", err)
+                    }
+                });
+            })
         });
-        createQueries.forEach((query, index) => {
-            dbPool.query(query);
-        })
         console.log("Remade Database");
     }
     console.log(rows);
